jesus christ, i fixed that sound bug from day 012 omg

+++ src/game.hpp
@@ -4,7 +4,7 @@
 
 namespace Game {
 	struct GameState {
-		f32 tSine;
+		f64 tSine;
 		u32 greenOffset;
 		u32 blueOffset;
 	};
